+++ src/financial/index.ts
@@ -1,0 +1,3 @@
+export * from "./loan";
+export * from "./compound";
+export * from "./investment";